Improve guessing function
Guessing with repeating letters will only mark match the number
of letters in the word to be guessed.

As an example, if the word is "apple" and the guess is "pasta",
only the first 'a' will be marked as correct but in the wrong position
and not both 'a's.

Letters in the correct position will have priority,
so if the word again is "apple" and the guess is "holly",
the second 'l' will be marked as in the correct position and the
first 'l' will be marked as not in the word.